#20
 モーダルの中身をpropsに変更
